refactor : rename FrameStatus -> FrameState
