+++ src/test/java/bowling/domain/BowlingGameTest.java
@@ -5,7 +5,7 @@ import org.junit.Test;
 
 import java.util.Arrays;
 
-import static bowling.enums.FrameStatus.*;
+import static bowling.enums.FrameState.*;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 public class BowlingGameTest {
