+++ src/test/java/bowling/domain/FrameTest.java
@@ -1,10 +1,9 @@
 package bowling.domain;
 
-import bowling.enums.FrameStatus;
 import org.junit.Before;
 import org.junit.Test;
 
-import static bowling.enums.FrameStatus.*;
+import static bowling.enums.FrameState.*;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class FrameTest {
